+++ examples/basic/data.rs
@@ -1,8 +1,13 @@
+use std::net::Ipv4Addr;
+
 use afire::{Content, HeaderType, Method, Query, Response, Server};
 
 use crate::Example;
 
-// Send data to server with a Query String, Path params and Form Data
+// You can run this example with `cargo run --example basic -- data`
+
+// In this example we will work with data in the request using query params and form data
+
 pub struct Data;
 
 impl Example for Data {
@@ -12,7 +17,7 @@ impl Example for Data {
 
     fn exec(&self) {
         // Create a new Server instance on localhost port 8080
-        let mut server = Server::<()>::new("localhost", 8080);
+        let mut server = Server::<()>::new(Ipv4Addr::LOCALHOST, 8080);
 
         // Define a route to handel query string
         // This will try to find a name value pair in the query string
@@ -33,7 +38,6 @@ impl Example for Data {
             // The body of requests is not part of the req.query
             // Instead it is part of the req.body but as a string
             // We will need to parse it get it as a query
-            // This is *super* easy to do with afire
             let body_data = Query::from_body(&String::from_utf8_lossy(&req.body));
 
             let name = body_data.get("name").unwrap_or("Nobody");
@@ -42,7 +46,7 @@ impl Example for Data {
             // Create a new response, with the following default data
             // - Status: 200
             // - Data: OK
-            // - Headers: Vec::new()
+            // - Headers: []
             Response::new()
                 // Set the response body to be text
                 .text(text)