+++ examples/basic/middleware.rs
@@ -1,24 +1,28 @@
 use afire::{
     middleware::{MiddleResult, Middleware},
-    Content, Method, Request, Response, Server,
+    Content, Header, Method, Request, Response, Server,
 };
 
 use crate::Example;
 
-// In afire Middleware is a trait that is implemented and can modify / overwrite Requests and Response
-// The Middleware functions for this are `pre` and `post` for before and after the routes
-// These functions will return a MiddleResponse / MiddleRequest which is an enum with the following
-//  - `Continue` to not affect the Request / Response
-//  - `Add(Request | Response)` to modify the Request / Response and continue to run Middleware (if any)
-//  - `Send(Response)` to send a Response to the client immediately
-// Just like routes, middleware is executed in reverse order that they are defined.
-// So the most recently defined middleware is executed first.
+// In afire Middleware is a trait that is implemented and can modify Requests and Response before and after Routes
+// You can use Middleware to Log Requests, Ratelimit Requests, add Analytics, etc.
+// The Middleware functions for this are `pre` and `post` for before and after the routes, there is also `end` which is called after the response is sent to the client
+//
+// There are two types of hooks: raw and non-raw.
+// The raw hooks are passed a Result, and their default implementation calls the non-raw hooks if the Result is Ok.
+// This allows you to handle errors (like page not found), while maintaining a clean API for middleware that doesn't need to handle errors.
+//
+// In the different middleware hooks you can return a MiddleResult, which is an enum with 3 variants:
+// - Continue: Continue to the next middleware or route
+// - Abort: Stop the middleware chain
+// - Send: Immediately send this response to the client and stop the middleware chain
+//
+// For more info, checkout the documentation for Middleware here: https://docs.rs/afire/latest/afire/middleware/trait.Middleware.html
 
-// You could use middleware to log requests, check if a user is logged in, Implement RateLimiting, add Analytics, etc.
-// In this example, we will make a very simple Request Logger
+// Lets make a Middleware that will log the request to the console
+// And to show how to modify the response, we will add a header to the response
 
-// First we define the Struct to Implement Middleware on
-// It can have values, but in this case that is not needed
 struct Log;
 
 // Now we will Implement Middleware for Log
@@ -28,11 +32,14 @@ impl Middleware for Log {
     fn pre(&self, req: &mut Request) -> MiddleResult {
         // Print some info
         println!("[{}] {} {}", req.address.ip(), req.method, req.path);
-        // Note: req.address also has the client port
-        // This is being removed with
-        // Ex: 127.0.0.1:6264 => 127.0.0.1
 
         // Continue to forward the request to the next middleware or route
+        MiddleResult::Continue
+    }
+
+    // Lets also modify the outgoing response by adding a header
+    fn post(&self, _req: &Request, res: &mut Response) -> MiddleResult {
+        res.headers.push(Header::new("X-Example", "Middleware"));
         MiddleResult::Continue
     }
 }
@@ -43,6 +50,7 @@ impl Example for MiddlewareExample {
     fn name(&self) -> &'static str {
         "middleware"
     }
+
     fn exec(&self) {
         // Create a new Server instance on localhost port 8080
         let mut server = Server::<()>::new("localhost", 8080);
@@ -57,7 +65,7 @@ impl Example for MiddlewareExample {
         Log.attach(&mut server);
 
         // You can now goto http://localhost:8080 you should see that the request is printed to the console
-        // It should look something like this: `[127.0.0.1:xxxxx] GET `
+        // It should look something like this: `[127.0.0.1] GET `
 
         // Start the server
         // This will block the current thread